backend/requirements.txt
fastapi
uvicorn
ollama
sentence-transformers
faiss-cpu
langchain-community